fix compiling errors, global variables SHOULD NOT be defined in header file
